Updated database code to sqlalchemy
