+++ requirements.txt
@@ -7,3 +7,4 @@ MarkupSafe==1.0
 Werkzeug==0.14.1
 WTForms==2.2.1
 Flask-Assets
+flask-sqlalchemy
