+++ webapp/__init__.py
@@ -1,27 +1,32 @@
 import os
 from flask import Flask
-from flask import redirect,url_for
+from flask import render_template
 from flask_assets import Environment, Bundle
+from flask_sqlalchemy import SQLAlchemy
 
-def create_app():
-    app = Flask(__name__)
-    # existing code omitted
-    assets = Environment(app)
-    assets.url = app.static_url_path
-    scss = Bundle('assets/scss/style.scss', filters='pyscss', output='css/style.css')
-    assets.register('scss_all', scss)
 
-    app.config.from_mapping(
-        SECRET_KEY='dev',
-        # DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
-        WTF_CSRF_SECRET_KEY="a csrf secret key",
-    )
+# def create_app():
+app = Flask(__name__)
+# existing code omitted
+assets = Environment(app)
+assets.url = app.static_url_path
+scss = Bundle('assets/scss/style.scss', filters='pyscss', output='css/style.css')
+assets.register('scss_all', scss)
 
-    from . import project
-    app.register_blueprint(project.bp)
+app.config.from_mapping(
+    SECRET_KEY='dev',
+    # DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
+    WTF_CSRF_SECRET_KEY="a csrf secret key",
+    SQLALCHEMY_TRACK_MODIFICATIONS=True,
+)
 
-    @app.route("/")
-    def hello():
-        return redirect(url_for('project.index'))
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+db = SQLAlchemy(app)
+db.create_all()
+from webapp import project
 
-    return app
+    # @app.route("/")
+    # def hello():
+    #     return render_template('index.html')
+    #
+    # return app
