+++ webapp/database.py
@@ -1,126 +1,35 @@
-import sqlite3
 from werkzeug.security import generate_password_hash
-
-def dict_factory(cursor, row):
-    d = {}
-    for idx, col in enumerate(cursor.description):
-        d[col[0]] = row[idx]
-    return d
-
-class sqllite():
-    def __init__(self):
-        self.conn = sqlite3.connect('database.db',check_same_thread=False)
-        self.conn.row_factory = dict_factory
-        print("Opened database successfully")
-
-    def create_table(self):
-        try:
-            self.conn.execute('CREATE TABLE conversation (input TEXT, output TEXT)')
-        except Exception as e:
-            print("Dataabase aleary exists")
-        finally:
-            print("Table created successfully")
-
-    def insert_data(self,input,output):
-        self.conn.execute('''INSERT INTO conversation(input,output) VALUES(?,?)''',(input,output))
-        self.conn.commit()
-        print("Inserted data to the database")
-
-    def show_data(self):
-        output = self.conn.execute('SELECT * FROM conversation')
-        # print(output.fetchall())
-
-    def get_output(self,input):
-        query = self.conn.execute('''SELECT output FROM conversation WHERE input=(?)''',(input,))
-        output = query.fetchall()
-        if output:
-            print(output)
-            # print(output[0][0])
-            return str(output[0]['output'])
-
-        else:
-            return "none"
-
-    def close(self):
-        self.conn.commit()
-        self.conn.close()
-
-class Authentication():
-    def __init__(self):
-        self.conn = sqlite3.connect('database.db',check_same_thread=False)
-        self.conn.row_factory = dict_factory
-        print("Opened database successfully")
-
-    def create_table(self):
-        try:
-            self.conn.execute('CREATE TABLE authentication (id INTEGER PRIMARY KEY AUTOINCREMENT,username TEXT, email TEXT,password TEXT)')
-        except Exception as e:
-            print("Dataabase aleary exists")
-        finally:
-            print("Table created successfully")
-
-    def create_account(self,username,email,password):
-        try:
-            self.conn.execute('''INSERT INTO authentication(username,email,password) VALUES(?,?,?)''',(username,email,generate_password_hash(password)))
-            self.conn.commit()
-            print("Inserted data to the database")
-            return 'success'
-        except Exception as e:
-            print(e)
-            return 'email'
-
-    def get_id(self,username):
-        output = self.conn.execute('SELECT id FROM authentication WHERE username = ?', (username,))
-        output = output.fetchone()
-        print("id {}".format(output))
-        return output
-
-    def get_user_from_name(self,username):
-        try:
-            output = self.conn.execute('SELECT * FROM authentication WHERE username = ?', (username,))
-            output = output.fetchone()
-            print("all the values {}".format(output))
-            return output
-        except Exception as e:
-            print(e)
-
-    def get_user_from_email(self,email):
-        try:
-            output = self.conn.execute('SELECT * FROM authentication WHERE email = ?', (email,))
-            output = output.fetchone()
-            print("info from email {}".format(output))
-            return output
-        except Exception as e:
-            print(e)
-
-    def get_user_from_id(self,id):
-        try:
-            output = self.conn.execute('SELECT * FROM authentication WHERE id = ?', (id,))
-            output = output.fetchone()
-            print("output from id {}".format(output))
-            return output
-        except Exception as e:
-            print(e)
-
-    def get_password(self,email):
-        try:
-            output = self.conn.execute(''' SELECT password FROM authentication WHERE email=(?) ''',(email,))
-            output = output.fetchall()
-            if output:
-                print(output)
-                return output
-        except Exception as e:
-            raise
+from webapp import db
 
 
 
-    def close(self):
-        self.conn.commit()
-        self.conn.close()
+class Conversation(db.Model):
+    id = db.Column(db.Integer,primary_key=True)
+    input = db.Column(db.String(300),unique=True,nullable=False)
+    output = db.Column(db.String(300),nullable=False)
 
-# sql = sqllite()
-# # # sql.create_table()
-# sql.insert_data("What can you do?","I can chat, ask me something")
-# # sql.show_data()
-# print(sql.get_output("how?"))
-# sql.close()
+    def add(chat):
+        db.session.add(chat)
+        db.session.commit()
+
+
+
+
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    password = db.Column(db.String(120), nullable=False)
+
+    def __init__(self,username,email,password):
+        self.username = username
+        self.password = generate_password_hash(password)
+        self.email = email
+
+    def create(user):
+        db.session.add(user)
+        db.session.commit()
+
+    def __repr__(self):
+        return '<User %r>' % self.username
