+++ webapp/route.py
@@ -1,0 +1,112 @@
+from flask import Blueprint,flash,request,g,session,url_for,redirect,render_template,jsonify
+
+from webapp.database import Conversation,User
+from wtforms import BooleanField,StringField,PasswordField,validators
+from flask_wtf import Form
+from werkzeug.security import check_password_hash, generate_password_hash
+import functools
+
+bp = Blueprint('project',__name__,url_prefix='/project')
+@bp.before_app_request
+def load_logged_in_user():
+    user_id = session.get('user_id')
+
+    if user_id is None:
+        g.user = None
+    else:
+        g.user = User.query.filter_by(id=user_id).first()
+
+
+@bp.route("/register",methods=['GET','POST'])
+def register():
+    if request.method == 'POST':
+        # auth = Authentication()
+        error = None
+        form = request.form
+
+        # return str(form)
+        output = User.query.filter_by(email=form['email']).first()
+        if output is not None:
+            print(output)
+            error = 'User {} is already registered.'.format(form['name'])
+        else:
+            user = User(username=form['name'],email=form['email'],password=form['password'])
+            User.create(user)
+            flash('Account created successfully')
+            return redirect(url_for('project.login'))
+
+
+        flash(error)
+    return render_template('register.html')
+
+@bp.route("forgot",methods=['GET','POST'])
+def forgot():
+    if request.method == 'POST':
+        error = None
+        error = "Not implimented"
+        flash(error)
+    return render_template('forgot.html')
+
+@bp.route("/login",methods=['GET','POST'])
+def login():
+    if request.method == 'POST':
+        error = None
+        # form = LoginForm(request.form)
+        form = request.form
+        user = User.query.filter_by(email=form['email']).first()
+        password = form['password']
+        if user is None:
+            error = 'Username or Email Id not found.'
+
+        elif not check_password_hash(user.password, password):
+            error = 'Incorrect password.'
+
+        if error is None:
+            session.clear()
+            session['user_id'] = user.id
+            return redirect(url_for('project.index'))
+
+        flash(error)
+    return render_template('login.html')
+
+@bp.route("/index")
+def index():
+    return render_template('index.html')
+
+@bp.route('/logout')
+def logout():
+    session.clear()
+    return redirect(url_for('project.index'))
+
+def login_required(view):
+    @functools.wraps(view)
+    def wrapped_view(**kwargs):
+        if g.user is None:
+            return redirect(url_for('project.login'))
+
+        return view(**kwargs)
+
+    return wrapped_view
+
+
+
+
+@bp.route("/support")
+def support():
+    return render_template('support.html')
+
+
+
+@bp.route("/gethint.php",methods=['GET', 'POST'])
+def getHint():
+    if request.method == 'GET':
+        msg = request.args.get('q')
+        response = Conversation.query.filter_by(input=msg.lower()).first()
+        if response:
+            return jsonify(response)
+        else:
+            # return ' '.join(matching)
+            return jsonify("I dont know what you are talking about")
+
+    elif request.method == 'POST':
+        return "me again"
